Move comparison operation into operators section.
